Directional scaling for drag and lift
Was explaining what I did and realised I messed up calculating lift. One was an old mistake: lift is just the player's up vector when it should be perpendicular to the flowDirection.

With the 50% Changed Drag commit I was adding a directional scale to drag. When you're moving in the direction you're facing you get less drag because air flowing past your front is more aerodynamic (it still worked for going backward which I needed to change).

I ended up applying that directional scale to lift as well because lift always has to be less than drag or it accelerates velocity to essentially infinity. It mostly work but still feels off, so I'll be fixing that.