independent directional scaling for drag and lift
I intend my flight to be pretty similar to games like The Falconeer, Voron, Sky: Children of Light and yours.

Getting in to the maths... I have four forces deciding the player's movement (rotation just snaps to the camera for now): thrust (applied for just one frame each time we flap, in the direction we're pointing), gravity (constant), lift (applied in the direction perpendicular to velocity's direction and the right direction), and drag (applied to the opposite direction of the velocity). I've been calculating lift and drag using their standard equations, ignoring everything except for velocity and the lift/drag coefficient.

After comparing and contrasting a spectrum of realistic to unrealistic flight in a few games, I realised the unrealistic ones seem like they're allowing the player to control the strength of their forces based on the direction they want to go. So instead of the coefficients with AoA and stall, I scale lift and drag off of how much they will help the velocity match the direction the player is facing / the camera direction (remember, rotation just snaps for now) / the player's forward vector.

This is done by finding the difference between the velocity and forward direction, getting the normalised vector (direction) of that difference, and then scaling the base drag/lift off of the the dot product of the difference direction and drag/lift's direction. Essentially, the greater the angle between the direction of drag/lift and this difference direction, the smaller the scaling gets 1 to -1. If they between angle is more than 90 degrees (approaching opposite) then I just clamp the scaling to a minimum value above 0.

It seems to be pretty close, with quite a few kinks that I think I can adjust for (like clamping opposites to a minimum instead of -1). Next, I'll probably do rotation since it's probably the reason for some of those kinks.