Smooth desired scaling
Lift and drag scale in magnitude based off of how much they can get the velocity direction to match the desired direction, but there was a problem. They could overshoot this value and that caused us to go back and forth between extreme values (0 multiplier when we overshot,  and 1 multiplier when we're below again). Fixed it by getting the maximum magnitude of lift we could apply before overshooting (projection stuff with desired difference).

Now gonna change how lift is generated in the first place